modify vocal acoustics to accomodate new outputs and separate modules
